day47_v1

+++ srcs/ft_printf.h
@@ -124,7 +124,7 @@ int							sum_decimal_const(t_float *a, \
 t_float						*power(\
 								unsigned long int base, short int power, \
 								t_float *exp);
-int							get_decimal(t_dbl dbl, t_float *decimal, \
+size_t						get_decimal(t_dbl dbl, t_float *decimal, \
 								short int fraction_length, short int *exponent);
 int							get_integer(t_dbl dbl, t_float *integer, \
 								short int *exponent);
